UBUNTU: SAUCE: ubuntu_ltp_controllers: remove source code to save disk space
When running on some smaller instances like the asus-pe100a (with 16G
SD card, but normally with 4G free space). It is better to have the LTP
source code removed to ensure the system won't experience the disk
out-of-space issue.

We're already doing this for ubuntu_ltp_syscalls test, and we just spot
similar problem for ubuntu_ltp_controllers with F/ein in 20250317.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>