UBUNTU: SAUCE: ubuntu_nvidia_server_driver: Fail the test if apt-cache search returns empty
apt-cache search --names-only "^linux-modules-nvidia-[0-9]+-
server(-open)?-$(uname -r)$" can return empty due to a number of
reasons such as -server variant not available, specific kernel
not supported, or some issue with the package archive. In case this
happens, the for loop will never be entered and test will pass by
default.

This commit checks if the `apt-cache search` returns empty, if so,
it fails the test.

Signed-off-by: Abdur Rahman <abdur.rahman@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Abdur Rahman <abdur.rahman@canonical.com>