UBUNTU: SAUCE: Purge docker.io after test for fan/docker smoke tests
KERNTT-1553

Certification team is asking us to not just remove docker.io after test,
but also purge it. Otherwise it will block their test on manually provisioned
MTK devices.

For the overlayfs, it needs to be fixed in the overlay-shiftfs-tests repo.

Only these three tests will install docker.io and be tested on MTK devices.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: jianhuilee <jianhui.lee@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>