UBUNTU: SAUCE: ubuntu_nvidia_fs: fixes for the test
We need to modprobe nvidia-fs so that it is loaded before the docker
runs the test. Otherwise the tests in docker are not able to find
nvidia-fs module.

No need to install nvidia drivers, as the drivers are installed by
our provisioning script.

Prevent prompt in case the gpg key is being overwritten. That may
happen in case the test is run again.

Signed-off-by: Abdur Rahman <abdur.rahman@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Abdur Rahman <abdur.rahman@canonical.com>