UBUNTU: SAUCE: ubuntu_sysdig_smoke_test: enable test for ARM64 on Jammy+
KERNTT-1483
In the past we have this test disabled on ARM64 becuase of LP: #1670643.
Experiment shows this ubuntu_sysdig_smoke_test works with ARM64 on J/N now.

But the module build is still failing with LP: #2015841 on X/B/F ARM64.
Since those are already ESM series, we can leave them disabled.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>