UBUNTU: SAUCE: ubuntu_performance_lkp: fix test build on Jammy ARM64
BugLink: https://bugs.launchpad.net/bugs/2119503

The commit we use for Jammy does not include a arch fix for non-x86 systems
(commit c96dfa6 "Fix RISC-V Debian/Ubuntu dependencies install" in lkp-test).
This will cause the test to install i386 package on ARM64 systems. And
consequently causing setup failure to the test.

This change has been tested with Jammy AMD64 and ARM64. It's working as
expected but it will cause test case name change and introduce a "not a valid
jobfile" failures to the following tests:
  * cassandra-cs-localhost-1SSD-300s-ext4-3.yaml
  * hackbench-100-socket-process-100%.yaml
  * hackbench-8-socket-process-800%.yaml
  * hackbench-perf-socket-process-1600%.yaml
  * schbench-100%.yaml

This issue will be addressed in LP: #2071705

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>