UBUNTU: SAUCE: do not explictly install gcc-multilib
KTNT-1674
The gcc-multilib package is not available on Questing. This is causing
test setup failures.

Installing the build-essential package should be enough, as it will
automatically pull in all the necessary dependencies. Therefore, I am
removing the gcc-multilib requirement, and add build-essential if the
test requires gcc.

I reviewed our test plan and only applied changes to the tests that
we are using.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Acked-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>