UBUNTU: SAUCE: install libc6-dev-i386 for -m32 related tests
KTNT-1674
With the gcc-multilib removed from the required package list. These tests:
  * ubuntu_32_on_64
  * ubuntu_cve_kernel
  * ubuntu_qrt_kernel_panic
  * ubuntu_qrt_kernel_security

Have been found failing with:
  fatal error: bits/libc-header-start.h: No such file or directory

This is because they will be compiled with a -m32 flag. Install libc6-dev-i386
to fix this issue. This package is available for AMD64 from Trusty to Questing.

ubuntu_32_on_64 is a AMD64 only test, therefore the arch check has been skipped.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Acked-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>