UBUNTU: SAUCE: ubuntu_bpf: add gcc-multilib for AMD64
After removing gcc-multilib from the requirement in commit "UBUNTU: SAUCE:
do not explicitly install gcc-multilib", we found that ubuntu_bpf on Bionic
will need libc6-dev-i386 too, this was not discovered during development
phase because the setup task didn't fail because of this.

Since installing libc6-dev-i386 will eventually pull in gcc-multilib, and
it is available for AMD64 from Trusty all the way to Resolute now, let just
add this package back for this test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Manuel Diewald <manuel.diewald@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>